converting doubles to floats, big ints to small ints

+++ include/superpixels/ConvexHullifier.h
@@ -11,25 +11,25 @@ class ConvexHullifier
 
         void setParams(const SuperpixelParams & params);
 
-        void run(std::vector<std::vector<double>> & centers,
+        void run(std::vector<std::vector<float>> & centers,
                     std::vector<int> & center_counts,
                     std::vector<std::vector<cv::Point>> & superpixels,
-                    std::vector<std::vector<Eigen::Vector2d>> & superpixel_projections,
-                    std::vector<std::vector<Eigen::Vector2d>> & superpixel_convex_hulls,
-                    std::vector<Eigen::Matrix3d> & superpixel_rotations,
+                    std::vector<std::vector<Eigen::Vector2f>> & superpixel_projections,
+                    std::vector<std::vector<Eigen::Vector2f>> & superpixel_convex_hulls,
+                    std::vector<Eigen::Matrix3f> & superpixel_rotations,
                     const cv::Mat & depth_image);
 
     private:
 
-        void convexHull(const std::vector<Eigen::Vector2d> & superpixel, std::vector<Eigen::Vector2d> & convex_hull);
+        void convexHull(const std::vector<Eigen::Vector2f> & superpixel, std::vector<Eigen::Vector2f> & convex_hull);
 
-        bool polarSort(const Eigen::Vector2d & a, const Eigen::Vector2d & b, const Eigen::Vector2d & lowest);
+        bool polarSort(const Eigen::Vector2f & a, const Eigen::Vector2f & b, const Eigen::Vector2f & lowest);
 
-        bool ccw(const Eigen::Vector2d & a, const Eigen::Vector2d & b, const Eigen::Vector2d & c);
+        bool ccw(const Eigen::Vector2f & a, const Eigen::Vector2f & b, const Eigen::Vector2f & c);
 
-        void grahamScan(const std::vector<Eigen::Vector2d> & superpixel, std::vector<Eigen::Vector2d> & convex_hull);
+        void grahamScan(const std::vector<Eigen::Vector2f> & superpixel, std::vector<Eigen::Vector2f> & convex_hull);
 
-        Eigen::Vector3d projectPointOntoPlane(const Eigen::Vector3d & regionPt);
+        Eigen::Vector3f projectPointOntoPlane(const Eigen::Vector3f & regionPt);
 
         SuperpixelParams params_;
